Expanded Customization Options
Broke out most of the timeline selectors as setting options to allow
the developer to more easily utilize the timeline functionality in
their page structure. This also required almost all the selector code
to be rewritten to utilize the settings.

Broke out “EX” as a setting to allow other Suffix to be used for
expanded content.

Added additional comments to provide guidance around where IDs and EX
IDs for timeline items need to be placed (in relation to the selectors).

Changed StartOpen to use ID instead of EX ID. This simplified the coded
needed to open the items.

Added oneOpen option which allows only one timeline item to be open at
a time. This is overridden by the clicking a Major (section) item.